UPDATE - Modulorizacion de funciones y archivos  terminada. Se corrigieron errores como ejecuta a secas y el enter
