--- .vscode/settings.json
@@ -1,8 +1,0 @@
-{
-    "files.associations": {
-        "stdio.h": "c",
-        "controlador.h": "c",
-        "globales.h": "c",
-        "string.h": "c"
-    }
-}
