+++ include/alu.h
@@ -5,6 +5,5 @@
 
 int aluGpo1(char *operacion, char *registro, int *valor);
 int aluGpo2(char *operacion, char *registro);
-int *obtenerRegistro(const char *nombre);
 
 #endif
