+++ include/comando.h
@@ -4,6 +4,7 @@
 #include "include/controlador.h"
 
 int extraerComando(buffer *bufferC);
+void leerComando(char *comando);
 void limpiarBuffer(buffer *bufferC);
 
 #endif // COMANDO_H
