+++ include/controlador.h
@@ -10,5 +10,8 @@
 #include "include/imprimir.h"
 #include "include/interprete.h"
 #include "include/comando.h"
+#include "include/manejador_archivos.h"
+#include "include/instrucciones.h"
+#include "include/validar.h"
 
 #endif
