+++ include/globales.h
@@ -8,9 +8,9 @@
 #define NUM_OPS_GPO1 5
 #define NUM_OPS_GPO2 2
 
-extern int reg_ax, reg_bx, reg_cx, reg_dx = 0;
-extern int reg_pc, reg_id = 0;
-extern char reg_proceso[16] = "",reg_ir[16] = "";
+extern int reg_ax, reg_bx, reg_cx, reg_dx;
+extern int reg_pc, reg_id;
+extern char reg_proceso[16], reg_ir[16], comando[TAMANIO_COMANDO];
 
 static const char *const REGISTROS[] = {"Ax", "Bx", "Cx", "Dx"};
 static const char *const OPERACIONES_GPO1[] = {"MOV", "ADD", "SUB", "MUL", "DIV"};
