+++ include/instrucciones.h
@@ -1,0 +1,7 @@
+#ifndef INSTRUCCIONES_H
+#define INSTRUCCIONES_H
+
+void analizadorGpo1(char *tipo_operacion, char *operandos);
+void analizadorGpo2(char *tipo_operacion, char *registro);
+
+#endif
