+++ include/manejador_archivos.h
@@ -1,0 +1,6 @@
+#ifndef MANEJADOR_ARCHIVOS_H
+#define MANEJADOR_ARCHIVOS_H
+
+int leerArchivo(char *nombre_archivo);
+
+#endif
