+++ include/validar.h
@@ -1,0 +1,9 @@
+#ifndef VALIDAR_H
+#define VALIDAR_H
+
+int validarRegistro(const char *registro);
+int tipoOperacion(const char *operacion);
+int esNumeroValido(const char *str);
+int *obtenerRegistro(const char *nombre);
+
+#endif
