+++ src/comando.c
@@ -10,25 +10,22 @@ int extraerComando(buffer *bufferC){    //ejecuta a.asm     o   salir
 
     strcpy(copia, bufferC->comandoCompleto);
     strcpy(bufferC->comando, strtok(copia, " "));   //ejecuta o salir 
-    token = strtok(NULL, " ");
+    token = strtok(NULL, " ");  //a.asm
 
     if(strcmp("salir", bufferC->comando) == 0){
         printf("Saliendo...\n");
         exit(0);
     }
     else if(strcmp("ejecuta", bufferC->comando) == 0){  //ejecuta
-        if(token == NULL){
+        if(token == NULL){ //revisa la parte de la derecha 
+            imprimirError("Falta especificar archivo");
             return -1;
         }
         strcpy(bufferC->argumento, token); // a.asm
         strcpy(reg_ir, "");
         strcpy(reg_proceso, bufferC->argumento);
 
-        if (bufferC->argumento == NULL){
-            imprimirError("Falta especificar archivo");
-            return -1;
-        }
-        else if (leerArchivo(bufferC->argumento) == -1){     //a.asm
+        if (leerArchivo(bufferC->argumento) == -1){     //a.asm
         reg_id--;
         strcpy(reg_proceso, "");
         }
@@ -36,6 +33,18 @@ int extraerComando(buffer *bufferC){    //ejecuta a.asm     o   salir
         imprimirError("Comando no reconocido");
         return -1;
     }
+}
+
+void leerComando(char *comando){
+    if (fgets(comando, TAMANIO_COMANDO, stdin) == NULL){
+        printf("\nError leyendo entrada\n");
+        comando[0] = '\0'; // Indica error
+        return;
+    }
+    comando[strcspn(comando, "\n")] = 0;
+    
+    interprete(comando);
+
 }
 
 void limpiarBuffer(buffer *bufferC) {
