+++ src/manejador_archivos.c
@@ -1,0 +1,49 @@
+#include "include/controlador.h"
+
+int leerArchivo(char *nombre_archivo){  //a.asm
+    reg_id++;
+    FILE *archivo = fopen(nombre_archivo, "r");
+    if (archivo == NULL){
+        imprimirError("Archivo no encontrado");
+        return -1;
+    }
+
+    char linea[TAMANIO_LINEA];
+    reg_pc = 1;     //Reinicio de PC
+    
+    imprimirTabla();
+    
+    while (fgets(linea, sizeof(linea), archivo) != NULL){   //linea = MOV Ax,7 o INC Ax
+        linea[strcspn(linea, "\n")] = 0;
+        
+        strcpy(reg_ir, linea);
+        
+        char copia[TAMANIO_LINEA];
+        strcpy(copia, linea);
+        char *token = strtok(copia, " ");       //token = MOV
+        
+        if (token == NULL){
+            continue;
+        }
+        
+        int tipo_op = tipoOperacion(token);     //1 o 2
+        
+        char *operandos = strtok(NULL, "");     //Ax,7
+        if (operandos == NULL){
+            imprimirFilaConError("Cantidad incorrecta de operandos");
+            continue;
+        }
+
+        if (tipo_op == 1){
+            analizadorGpo1(token, operandos);   //MOV y Ax,7
+        } else if (tipo_op == 2){
+            analizadorGpo2(token, operandos);
+        } else{
+            imprimirFilaConError("Instrucción no reconocida");
+            continue;
+        }
+    }
+
+    fclose(archivo);
+    return 0;
+}
